embedding_generator/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
PyYAML>=6.0

# FAISS for similarity search
faiss-cpu>=1.7.4
# Use faiss-gpu instead if you have CUDA support:
# faiss-gpu>=1.7.4

# Data processing
jsonlines>=3.0.0
datasets>=2.0.0

# Utilities
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: For advanced text preprocessing
nltk>=3.7
spacy>=3.4.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950